feat: add inventory and item logic (#18)
* feat: added item and inventory classes

This commit adds barebones classes for the item and inventory systems, to be expanded later

* feat: Add logic for picking up items

This commit adds logic for adding items to inventory, as well as something that happens when you pick up an item that you already have.

* feat: added stackable distinction

This commit adds logic for stackable items, where some items can be stacked in the same slot (like consumables) while others are unique

* feat: updated TODO's

I added some TODO comments to the Inventory class

* feat: added icons and max inventory size

This commit adds a folder, and 2 placeholder icons. It also adds a max inventory size for the inventory

* fix: updated TODO's

updated todo in inventory class

* Merged main into inventory hotbar

* Display logic + summaries

Added testitem prefab
added display logic for items
updated summaries

* Final functionality

Items are visible ingame now, and item uses are fixed

* Removed final Debug.Logs

removed comments, changed some summaries and removed the contents of the testfunctions

* Removed Iconswitcher and added summary

Removed redundant script and added a summary instead of comments in the item class

* Folder restructuring

Moved scripts of inventory and renamed to folder to be more logical

* Added input for inventory

Added a comment to inputmanager and added controls for adding and activating items. Also added objects and ui to all relevant scenes

* Removed icons and formatted

Removed copyrighted icons, formatted inventory class

* Bugfix itemswitch

Only switch items if queue is not empty

* Anchored item position and better object list structure

Anchored the inventory display so it works on any resolution and rearranged object list

* Added more descriptions for parameters in inventory class

* Added doorcollision scene and class and added door object

* Added tempitemID class, added door logic, added summaries, added door prefab

* Changed door texture

* Add doors?

* Add doors?

* fix: summary formatting

this commit will fix some of the formatting for this sunmmary

Co-authored-by: Luka <51298607+LuukBerkers@users.noreply.github.com>

* fix: tag efficiency

this commit increases tag checking efficiency

Co-authored-by: Luka <51298607+LuukBerkers@users.noreply.github.com>

* Update Inventory.cs

* fix: extra whiteline

Co-authored-by: Luka <51298607+LuukBerkers@users.noreply.github.com>

* fix: removed unused references

* feat: changed comments to summaries on rawimage objects

this comment adds a summary to 2 things

* feat: removed test prefabs, functions and scripts

This commit will remove all the things I used for testing the codes functionality. Insctructions for how to remake them have been added to pr description

* fix: item newline

* fix: hardcode and style issues item/inventory

This commit will fix the last few style issues and 1 piece of hardcoding in the inventory

* fix: style changes inventory and item

This commit will use editorconfig to clean up some style issues

* fix: whitelines

* fix: naming conventions

* fix: naming conventions

* feat: more xml documentation

this commit will add documentation for 2 extra parameters

* fix: id in editor

this commit will fix door collision

* fix: update door prefab

this commit will update door prefab

* fix: save scene

this commit will fix door collision

* fix: summary and naming convention

summary for parent and doorid and removed underscore

* fix: typo

* fix: remove whiteline

* fix: code optimization and comments

Some changes to make code faster, and some comments added and changed for clarity

* feat: empty inventory icon and naming

this comment adds a new texture, and changes some names

* fix: image name and inventory size

* fix: delete empty items folder

* fix: test scene

This commit will fix to the final version of the scene

* Update aplib.net-demo/Assets/Scripts/InputManager.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Item.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/InputManager.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Item.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Item.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Item.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

* fix: comments, reset

this comment will change some syntax, removes some comments, and some needless functionality

* fix: comment

* Update aplib.net-demo/Assets/Scripts/Inventory/Inventory.cs

Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>

---------

Co-authored-by: Pepijn Lamers <p.p.lamers@students.uu.nl>
Co-authored-by: Luka <51298607+LuukBerkers@users.noreply.github.com>
Co-authored-by: Quake <74355598+QuakeEye@users.noreply.github.com>