feat: add ammo item (#67)
* Add AmmoPouch class for ammunition management

Implemented a new class, AmmoPouch, which handles the player's ammunition inventory. It keeps track of the current and maximum ammunition count, allows adding ammunition up to the set maximum, and provides a method for using ammo when available.

* feat: Refactor inventory system and add ammo pouch support

The inventory system in Inventory.cs has been considerably refactored. Specifically, an 'AmmoPouch' component is now referenced and is used for handling the pickup of AmmunitionItems. The item pickup method has also been refactored for clarity and efficiency, notably improving support for non-stackable items.

* feat: Add ammo prefab and corresponding assets for ammo pickup scene.

This commit includes the creation of an ammo prefab in Unity and the associated models for use within the game. The prefab contains the necessary components and parameters, including a script for handling the functionality of the ammo within the game. An 'ammo_sniper.fbx' file is also added, which holds the 3D model for the newly added ammo.

* chore: change guid of pickupable items

Due to a change on my side, the guid of the pickupable items was changed.

* Conform to code review

* fix: fix typo